More clean up of v12 and v13
